fix(auto-memory): address reconcile orchestration review feedback
Run absence/tombstone advancement once per batch reconcile, extract
ingest into auto_memory_ingest.py to break the reconcile import cycle,
and add absence_confirm_polls coverage for reconcile_sources.
